agent/requirements.txt
psutil
requests
pydivert>=2.1.0
pystray>=0.19.0
Pillow>=10.0.0
